Extract code into 'Eval' module
We extract code related to evaluation from the `Core` module and move it
to a new `Eval` module. This diminishes the size of the large core
module and will enable us to split the `Core` module up even further.
